ai/requirements.txt
fastapi
uvicorn
openai
anthropic
google-generativeai
redis
huggingface_hub
transformers
tiktoken
pydantic
pydantic-settings
python-dotenv
asyncio
locust
